Code split, applied review request changes
